add new java files

+++ lectures/code/threads/Account.java
@@ -1,5 +1,5 @@
 class Account {
-	float balance = 0.0;
+	double balance = 0.0;
 	public void deposit(float value) {
 		balance = balance + value;
 	}
